requirements.txt
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
exceptiongroup==1.1.1
idna==3.4
iniconfig==2.0.0
jsonschema==4.17.3
packaging==23.0
paho-mqtt==1.6.1
pluggy==1.0.0
psycopg2-binary==2.9.5
pvporcupine==2.1.4
pvrecorder==1.1.1
PyAudio==0.2.13
pyproject-toml==0.0.10
pyrsistent==0.19.3
pytest==7.2.2
python-decouple==3.8
python-dotenv==1.0.0
requests==2.28.2
SpeechRecognition==3.9.0
toml==0.10.2
tomli==2.0.1
urllib3==1.26.14
